refactor: Phase 3 — CLI God Object 분해 (session_state + cmd_schedule 추출)
_ResultCache + 세션 상태 → session_state.py (~115줄 추출)
cmd_schedule + 헬퍼 2개 → cmd_schedule.py (~220줄 추출)
re-export로 하위 호환 유지. 테스트 mock 경로 갱신.
행동 변경 없음.

품질 게이트: lint 0, format clean, type 0, 3224 passed.
